nsstrings: add NSStringToString
Closes #438
